database master-slave

+++ app/models/comment.py
@@ -1,15 +1,16 @@
-from sqlalchemy import Column, BigInteger, Text, ForeignKey, DateTime, func
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Text, func
 from sqlalchemy.orm import relationship
+
 from app.core.db import Base
 
 
 class Comment(Base):
     __tablename__ = "comments"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     content = Column(Text, nullable=False)
-    post_id = Column(BigInteger, ForeignKey("posts.id"), nullable=False)
-    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)
+    post_id = Column(Integer, ForeignKey("posts.id"), nullable=False)
+    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
     post = relationship("Post", back_populates="comments")
